Move fixtures to dedicated subfolder
